+++ gfx/video_thread_wrapper.h
@@ -67,7 +67,7 @@ enum thread_cmd
    CMD_DUMMY = INT_MAX
 };
 
-typedef int (*custom_command_method_t)(void*);
+typedef uintptr_t (*custom_command_method_t)(void*);
 
 typedef bool (*custom_font_command_method_t)(const void **font_driver,
       void **font_handle, void *video_data, const char *font_path,
@@ -140,7 +140,7 @@ typedef struct thread_packet
       {
          custom_command_method_t method;
          void* data;
-         int return_value;
+         uintptr_t return_value;
       } custom_command;
 
       struct
@@ -272,7 +272,7 @@ bool video_thread_font_init(
       custom_font_command_method_t func,
       bool is_threaded);
 
-unsigned video_thread_texture_handle(void *data,
+uintptr_t video_thread_texture_handle(void *data,
       custom_command_method_t func);
 
 /* Barrier: wait until the video thread is idle (no pending frame).
